+++ client/keystore/keys.go
