+++ client/keystore/keys_test.go
