+++ client/keystore/keystore.go
