+++ client/lite/statik/statik.go
