+++ cmd/iris/main.go
@@ -6,8 +6,8 @@ import (
 	"github.com/cosmos/cosmos-sdk/server"
 	svrcmd "github.com/cosmos/cosmos-sdk/server/cmd"
 
+	_ "github.com/irisnet/irishub/v2/client/lite/statik"
 	"github.com/irisnet/irishub/v2/cmd/iris/cmd"
-	_ "github.com/irisnet/irishub/v2/lite/statik"
 	"github.com/irisnet/irishub/v2/types"
 )
