+++ types/address.go
@@ -1,4 +1,4 @@
-package address
+package types
 
 import sdk "github.com/cosmos/cosmos-sdk/types"
